+++ lib/t/+mp/dme_branch_node_test.m
@@ -34,7 +34,7 @@ classdef dme_branch_node_test < mp.dme_branch_opf
             %% get bus mapping info
             obj.nbet = length(obj.cxn_type);
             for k = obj.nbet:-1:1
-                if dm.elements.is_index_name(obj.cxn_type{k})
+                if dm.elements.has_name(obj.cxn_type{k})
                     bus_dme{k} = dm.elements.(obj.cxn_type{k});
                     b2i_k{k} = bus_dme{k}.ID2i; %% bus element num to idx mapping
                 else
@@ -72,7 +72,7 @@ classdef dme_branch_node_test < mp.dme_branch_opf
         function obj = update_status(obj, dm)
             %% get bus status info
             for k = 1:obj.nbet
-                if dm.elements.is_index_name(obj.cxn_type{k})
+                if dm.elements.has_name(obj.cxn_type{k})
                     bus_dme = dm.elements.(obj.cxn_type{k});
                     bs = bus_dme.tab.status;    %% bus element status