+++ lib/t/t_mp_mapped_array.m
@@ -47,12 +47,12 @@ t_is(length(A), 4, 12, [t 'length']);
 t_ok(isequal(A.p_.vals, vals), [t 'vals']);
 t_ok(isequal(A.p_.names, names), [t 'names']);
 
-t = 'is_index_name : ';
-t_ok(A.is_index_name('vector'), [t 'vector']);
-t_ok(A.is_index_name('cellstr'), [t 'cellstr']);
-t_ok(A.is_index_name('struct'), [t 'struct']);
-t_ok(A.is_index_name('object'), [t 'object']);
-t_ok(~A.is_index_name('random'), [t 'random']);
+t = 'has_name : ';
+t_ok(A.has_name('vector'), [t 'vector']);
+t_ok(A.has_name('cellstr'), [t 'cellstr']);
+t_ok(A.has_name('struct'), [t 'struct']);
+t_ok(A.has_name('object'), [t 'object']);
+t_ok(~A.has_name('random'), [t 'random']);
 
 t = 'name2idx : ';
 t_is(A.name2idx('vector'), 1, 12, [t 'vector']);